data/pyproject.toml
[project]
name = "ai-interior-data"
version = "0.1.0"
description = "AI Home Styling Platform - Data Scripts"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "crawl4ai>=0.8.0",
]
